% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calculate_sentiment}
\alias{calculate_sentiment}
\title{Calculates the sentiment of each string.}
\usage{
calculate_sentiment(strings)
}
\arguments{
\item{strings}{A list of strings.}
}
\value{
An integer list of the sentiment scores of each string.
}
\description{
Calculates the sentiment of each string.
}
