% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_articles.R
\docType{data}
\name{cnn_articles}
\alias{cnn_articles}
\title{A dataset of CNN articles.}
\format{A data frame with 4 variables:
\describe{
 \item{title}{The title of the article.}
 \item{text}{The text in the article body.}
 \item{score}{The clickbait score of the article.}
 \item{sentiment}{The sentiment score of the article.}
 }}
\usage{
cnn_articles
}
\description{
A dataset containing the title and text of some CNN articles,
as well as some precalculated scores.
}
\keyword{datasets}
