% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{count_sig_words}
\alias{count_sig_words}
\title{Counts the number of significant words that appear in an article.}
\usage{
count_sig_words(sig_words, articles)
}
\arguments{
\item{sig_words}{A determined list of signicant words for each article.}

\item{articles}{A list of articles.}
}
\value{
A list of the number of signicant words in each article.
}
\description{
Counts the number of significant words that appear in an article.
}
