% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_sig_words}
\alias{get_sig_words}
\title{Gets the significant words from a string.}
\usage{
get_sig_words(titles)
}
\arguments{
\item{titles}{A list of strings.}
}
\value{
A list of signifcant words correpsonding to each string.
}
\description{
Gets the significant words from a string.
}
