% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{keyword_count}
\alias{keyword_count}
\title{Counts the occurences of keywords from the keyword vector for each item.}
\usage{
keyword_count(titles)
}
\arguments{
\item{titles}{A list of character vectors.}
}
\value{
A double list with the keyword count for each item.
}
\description{
Counts the occurences of keywords from the keyword vector for each item.
}
