% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Regular expressions that are likely to evoke strong emotions.}
\format{A character vector with 143 regular expressions:}
\source{
\url{http://nativeadvertising.com/contextwords/}
}
\usage{
keywords
}
\description{
A data set containing 143 regular expressions known to evoke
strong emotions.
}
\keyword{datasets}
