% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nypost_articles.R
\docType{data}
\name{nypost_articles}
\alias{nypost_articles}
\title{A dataset of New York Post articles.}
\format{A data frame with 4 variables:
\describe{
 \item{title}{The title of the article.}
 \item{text}{The text in the article body.}
 \item{score}{The clickbait score of the article.}
 \item{sentiment}{The sentiment score of the article.}
 }}
\usage{
nypost_articles
}
\description{
A dataset containing the title and text of some NY Post articles,
as well as some precalculated scores.
}
\keyword{datasets}
