% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{parse_punctuation}
\alias{parse_punctuation}
\title{Removes punctuation from words in a VCorpus object.}
\usage{
parse_punctuation(strings)
}
\arguments{
\item{strings}{A list of VCorpus objects.}
}
\value{
A list of character vectors stripped of punctuation.
}
\description{
Removes punctuation from words in a VCorpus object.
}
