% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{remove_stop_words}
\alias{remove_stop_words}
\title{Removes stop words from VCorpus objects.}
\usage{
remove_stop_words(articles)
}
\arguments{
\item{articles}{A list of vcorpus objects.}
}
\value{
VCorpus objects with the stop words removed from them.
}
\description{
Removes stop words from VCorpus objects.
}
