% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{tokenize}
\alias{tokenize}
\title{Tokenizes articles.}
\usage{
tokenize(articles)
}
\arguments{
\item{articles}{A list of articles.}
}
\value{
A list of VCorpus objects based on the articles.
}
\description{
Tokenizes articles.
}
